[#11] refactor: extract reference image reading into src/input-image.ts
